+++ app/src/main/AndroidManifest.xml
@@ -16,6 +16,9 @@
         android:usesCleartextTraffic="true"
         tools:targetApi="31">
         <activity
+            android:name=".buotg.view.user_group.StudyGroupActivity"
+            android:exported="false" />
+        <activity
             android:name=".buotg.view.user_setting.SettingActivity"
             android:exported="false" />
         <activity
