+++ app/src/main/java/com/cs501/cs501app/MainActivity.kt
@@ -34,6 +34,7 @@ import com.cs501.cs501app.buotg.database.entities.KVEntry
 import com.cs501.cs501app.buotg.database.repositories.AppRepository
 import com.cs501.cs501app.buotg.view.user_setting.SettingActivity
 import com.cs501.cs501app.buotg.view.user_setup.SetupActivity
+import com.cs501.cs501app.buotg.view.user_group.StudyGroupActivity
 import com.cs501.cs501app.example.WebViewDemo
 import com.cs501.cs501app.utils.GenericTopAppBar
 import com.cs501.cs501app.utils.TAlert
@@ -46,6 +47,7 @@ class MainActivity : AppCompatActivity() {
         WebViewDemo::class,
         SetupActivity::class,
         SettingActivity::class,
+        StudyGroupActivity::class,
 //        Calc1Activity::class,
 //        Calc2Activity::class,
 //        GeoQuizActivity::class,
